# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "轻量级文件上传与下载服务器，支持 zip/7z 自动解压",
        "title": "GFile API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "返回包含上传表单和文件列表的 HTML 页面",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "页面"
                ],
                "summary": "首页",
                "responses": {
                    "200": {
                        "description": "HTML 页面",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/download/{filename}": {
            "get": {
                "description": "以附件形式下载指定文件",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件操作"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文件内容",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "文件不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/files": {
            "get": {
                "description": "以 JSON 格式返回所有已上传文件的信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件操作"
                ],
                "summary": "文件列表",
                "responses": {
                    "200": {
                        "description": "文件列表",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "上传单个文件。支持普通文件以及 .zip / .7z 压缩包（自动解压）",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件操作"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "上传成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "文件已存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/view/{filename}": {
            "get": {
                "description": "在浏览器中内联展示文件内容（图片、文本等支持直接预览）",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件操作"
                ],
                "summary": "预览文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文件内容",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "文件不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}